# backend/openapi.yaml
openapi: 3.0.3
info:
  title: CYOPS Vulnerability Management API
  description: API endpoints for vulnerability tracking and management
  version: 1.0.0
  contact:
    name: CYOPS Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.cyops.com/api/v1
    description: Production
...
paths:
...
  /vulnerabilities/{id}/affected-systems:
    get:
      tags:
        - Affected Systems
      summary: Get affected systems for vulnerability
      description: Retrieve list of systems affected by a specific vulnerability
      operationId: getAffectedSystems
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Vulnerability UUID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AffectedSystem"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"

    post:
      tags:
        - Affected Systems
      summary: Add affected systems to vulnerability
      description: Associate one or more systems with a vulnerability
      operationId: addAffectedSystems
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Vulnerability UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - system_ids
              properties:
                system_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                  description: Array of affected system IDs
      responses:
        "200":
          description: Systems added successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VulnerabilityResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /affected-systems:
    get:
      tags:
        - Affected Systems
      summary: List all affected systems
      description: Retrieve paginated list of systems in the organization
      operationId: listAffectedSystems
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: systemType
          in: query
          description: Filter by system type
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AffectedSystem"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"

    post:
      tags:
        - Affected Systems
      summary: Create affected system
      description: Register a new system in the inventory
      operationId: createAffectedSystem
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAffectedSystemRequest"
      responses:
        "201":
          description: System created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/AffectedSystem"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Vulnerability:
      type: object
      required:
        - id
        - title
        - description
        - severity
        - status
        - discovery_date
        - created_by_id
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 255
        description:
          type: string
        severity:
          type: string
          enum: [CRITICAL, HIGH, MEDIUM, LOW, NONE]
        cvss_score:
          type: number
          format: float
          minimum: 0.0
          maximum: 10.0
          nullable: true
        cvss_vector:
          type: string
          maxLength: 100
          nullable: true
        cve_id:
          type: string
          pattern: "^CVE-[0-9]{4}-[0-9]{4,}$"
          nullable: true
        status:
          type: string
          enum: [OPEN, IN_PROGRESS, RESOLVED, VERIFIED, CLOSED, FALSE_POSITIVE]
        discovery_date:
          type: string
          format: date
        remediation_notes:
          type: string
          nullable: true
        impact_assessment:
          type: string
          nullable: true
        steps_to_reproduce:
          type: string
          nullable: true
        mitigation_recommendations:
          type: string
          nullable: true
        created_by_id:
          type: string
          format: uuid
        assigned_to_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    AffectedSystem:
      type: object
      required:
        - id
        - system_type
      properties:
        id:
          type: string
          format: uuid
        hostname:
          type: string
          nullable: true
        ip_address:
          type: string
          nullable: true
        asset_id:
          type: string
          nullable: true
        system_type:
          type: string
          enum:
            [
              SERVER,
              WORKSTATION,
              NETWORK_DEVICE,
              APPLICATION,
              CONTAINER,
              CLOUD_SERVICE,
              OTHER,
            ]
        description:
          type: string
          nullable: true
        environment:
          type: string
          enum: [PRODUCTION, STAGING, DEVELOPMENT, TEST]
          default: PRODUCTION
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateAffectedSystemRequest:
      type: object
      required:
        - system_type
      properties:
        hostname:
          type: string
        ip_address:
          type: string
        asset_id:
          type: string
        system_type:
          type: string
          enum:
            [
              SERVER,
              WORKSTATION,
              NETWORK_DEVICE,
              APPLICATION,
              CONTAINER,
              CLOUD_SERVICE,
              OTHER,
            ]
        description:
          type: string
        environment:
          type: string
          enum: [PRODUCTION, STAGING, DEVELOPMENT, TEST]
          default: PRODUCTION
    VulnerabilityResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Vulnerability"
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
    BadRequestError:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnauthorizedError:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ForbiddenError:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...